Update and rename use-chat-visibility.ts to updated_use-chat-visibility.ts
This updated version includes:

toggleVisibilityType for toggling between 'private' and 'public' visibility.

resetVisibilityType for resetting visibility to the initial state.

batchUpdateVisibility for updating the visibility of multiple chats at once.

Persistence of visibility settings to local storage.

getPublicChats for fetching all public chats.
